apis/v1: Split Probe types in their own file
