+++ pkg/apis/monitoring/v1/types_test.go
@@ -307,51 +307,3 @@ func TestValidateAuthorization(t *testing.T) {
 		})
 	}
 }
-
-func TestValidateProbeTargets(t *testing.T) {
-	tests := []struct {
-		name         string
-		probeTargets ProbeTargets
-		wantErr      bool
-	}{
-
-		{
-			name: "probe with static config target",
-			probeTargets: ProbeTargets{
-				StaticConfig: &ProbeTargetStaticConfig{
-					Targets: []string{"/probe"},
-					Labels:  map[string]string{"app": "foo"},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "probe with ingress target",
-			probeTargets: ProbeTargets{
-				Ingress: &ProbeTargetIngress{
-					Selector: metav1.LabelSelector{
-						MatchLabels: map[string]string{
-							"app": "foo",
-						},
-					},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "one of staticConfig and ingress is required",
-			probeTargets: ProbeTargets{
-				StaticConfig: nil,
-				Ingress:      nil,
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := tt.probeTargets.Validate(); (err != nil) != tt.wantErr {
-				t.Errorf("Validate() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
